Refactor imports in ComboCategories and ComboCategoryTypes components to use updated ComboboxItem type from unitTypes
